requirements.txt
pandas==2.3.1
yfinance==0.2.65
matplotlib==3.10.0
tabulate==0.9.0
riskfolio-lib==7.0.1